refactor + clean

+++ src/main.cpp
@@ -25,7 +25,7 @@ void run_experiment(input &par) {
 
         auto end_time = chrono::high_resolution_clock::now();
         auto duration = chrono::duration_cast<chrono::milliseconds>(end_time - start_time);
-        out.running_time = duration.count() / 1e+3;
+        out.running_time = static_cast<double>(duration.count()) / 1e+3;
 
         //cout << out << endl;
